+++ CHANGES.md
@@ -61,6 +61,7 @@ Changes in 13.0
 - Add AddonMakeUpdateAvailable() function in Addon.fnc.php
 - Check for add-on updates in Modules.inc.php & Plugins.inc.php
 - Add link to rosariosis.org/modules & rosariosis.org/plugins in Modules.inc.php & Plugins.inc.php
+- Security fix check sibling belongs to the authenticated parent in Registration.php & Registration.fnc.php, thanks to Kai Aizen
 
 Changes in 12.9.3
 -----------------